finished fixed houmwork_12_1 and fix readme
