+++ README.md
@@ -173,14 +173,14 @@
 
 Функция фильтрует транзакции по заданной валюте и возвращает итератор, который последовательно выдает только те транзакции, где валюта соответствует заданному значению.
 
-*   *Параметры:*
+*   **Параметры:**
     *   transactions (List[Dict[str, Any]]): Список словарей с транзакциями.
     *   currency (str): Код валюты для фильтрации, например, "USD".
 
-*   *Возвращаемое значение:*
+*   **Возвращаемое значение:**
     *   (Iterator[Dict[str, Any]]): Итератор с отфильтрованными транзакциями.
 
-*   *Пример:*
+*   **Пример:**
 
     ```python
     from your_module import filter_by_currency # Замените your_module на имя вашего модуля
@@ -196,13 +196,13 @@
 
 Функция генерирует описания транзакций из списка словарей с транзакциями.
 
-*   *Параметры:*
+*   **Параметры:**
     *   transactions (List[Dict[str, Any]]): Список словарей, представляющих транзакции.
 
-*   *Возвращаемое значение:*
+*   **Возвращаемое значение:**
     *   (Iterator[str]): Итератор описаний транзакций.
 
-*   *Пример:*
+*   **Пример:**
 
     ```python
     from your_module import transaction_descriptions # Замените your_module на имя вашего модуля
@@ -219,14 +219,14 @@
 
 Генератор, который создает номера банковских карт в формате XXXX XXXX XXXX XXXX в заданном диапазоне.
 
-*   *Параметры:*
+*   **Параметры:**
     *   start (int): Начальное значение диапазона (включительно).
     *   end (int): Конечное значение диапазона (включительно).
 
-*   *Возвращаемое значение:*
+*   **Возвращаемое значение:**
     *   (Iterator[str]): Итератор с номерами банковских карт в заданном диапазоне.
 
-*   *Пример:*
+*   **Пример:**
 
     ```python
     from your_module import card_number_generator # Замените your_module на имя вашего модуля
@@ -238,7 +238,7 @@
         0000 0000 0000 0004
         0000 0000 0000 0005
 
-### 5. Модуль `decorators.py` содержит функции:
+### 5. Модуль `decorators.py` содержит декоратор:
 
 ### `log(filename="log.txt")`
 
@@ -266,29 +266,27 @@
 **Примеры использования:**
 
 
-### Пример 1: Логирование в консоль
+*   **Пример 1: Логирование в консоль**
 
-```python
-@log()
-def my_func(x, y):
-    """Складывает два числа в консоль."""
-    return x + y
+     ```python
+    from typing import Any, Callable, Optional
+    @log()
+    def my_func(x, y):
+        """Складывает два числа в консоль."""
+        return x + y
+    print(my_func(5, 3)
+    >>> 8
 
-print(my_func(5, 3)
->>> 8
-```
+*   **Пример 2: Логирование в файл "my_log.txt"**
 
-### Пример 2: Логирование в файл "my_log.txt"
-
-```python
-@log(filename="my_log.txt")
-def my_func(x, y):
-    """Складывает два числа в файл."""
-    return X + Y
-
-print(my_func(5, 3)
->>> добавляется файл log.txt с значением "8"
-```
+    ```python
+    from typing import Any, Callable, Optional
+    @log(filename="my_log.txt")
+    def my_func(x: int, y: int) -> int:
+        """Складывает два числа в файл."""
+        return X + Y
+    my_func(5, 3)
+    >>> добавляется файл log.txt с значением "8"
 
 **Как это работает (пример с выводом в консоль):**
 
@@ -309,7 +307,66 @@ print(my_func(5, 3)
 
 Этот код предоставляет готовую основу для логирования, которую можно легко адаптировать.
 
+### 5. Модуль `utils.py` содержит функции:
 
+### `reading_json_file((file_path: str)`
+
+Функция принимает на вход путь до JSON-файла и возвращает список словарей с данными о финансовых транзакциях.
+
+*   **Параметры**
+
+    *   `file_path (str):` Путь до файла .json (по умолчанию в этом проекте `'../data/operations.json'`).
+
+*   **Возвращаемое значение**
+
+    *   `list[dict]:`  Cписок словарей с данными о финансовых транзакциях   
+
+### 5. Модуль `external_api.py` содержит функцию:
+
+### `convert_currency(currency: str, amount: float)`
+
+*   **Параметры:**
+    *   currency (str): Транзакция.
+    *   amount (float): Сумма, которую надо конвертировать.
+
+*   **Возвращаемое значение:**
+    *   `float:` Cумма транзакции.
+
+*   **Пример:**
+
+    ```python
+    import os
+    import requests
+    from dotenv import load_dotenv
+    print(convert_currency("USD", 100.0))
+    >>> 7820.3993 # Эта сумма конвертирования валюты на 30.06.2025г. 
+    
+### `convert_to_rub(operation: dict)`
+
+Функция конвертирования в рубли.
+
+*   **Параметры**
+
+    *   `operation: dict:` Словарь с данными о транзакции.
+
+*   **Возвращаемое значение**
+
+    *   `float:`  возвращает сумму транзакции (ключ amount) в рублях, тип данных float
+
+*   **Пример:**
+
+    ```python
+    import os
+    import requests
+    from dotenv import load_dotenv
+    print(convert_to_rub({
+    "id": 27192367, "state": "CANCELED",
+    "date": "2018-12-24T20:16:18.819037",
+    "operationAmount": {"amount": "991.49",
+    "currency": {"name": "руб.",
+    "code": "RUB"}}}))
+    >>> 991.49
+    
 ## Тестирование
 
 Для обеспечения надежности и качества вашего кода, были разработаны тесты с использованием pytest.
@@ -324,13 +381,20 @@ print(my_func(5, 3)
 
 1. Перейдите в каталог с вашим проектом в терминале.
 2. Запустите все тесты, выполнив команду:
-    ```
+   
+3. ```
    pytest -v --cov=your_module --cov-report=term-missing
    ```
+   
+   ```
+   pytest --cov=src --cov-report=html
+   ```
+
 Заменив `your_module` на имя вашего модуля Python (без расширения `.py`).
 * `-v`: (verbose) - отображает подробную информацию о прохождении тестов.
 * `--cov=your_module`: Указывает `pytest-cov` анализировать покрытие кода для вашего модуля.
 * `--cov-report=term-missing`: Отображает строки кода, которые не были покрыты тестами.
+* `--cov=src --cov-report=html`: Создаёт файл HTMLCOV в корне проекта со всеми параметрами тестов.
 
 ### Анализ результатов
