+++ log.txt
@@ -64,3 +64,4 @@ my_func - ok - 5
 my_func - ok - 5
 my_func - ok - 5
 my_func - ok - 5
+my_func - ok - 5
