+++ tests/test.txt
@@ -35,3 +35,4 @@ my_func - ok - 5
 my_func - ok - 5
 my_func - ok - 5
 my_func - ok - 5
+my_func - ok - 5
